move react sub-routers into top level app.js
